Move slurm functions into slurm.py
